Added JRuby benchmarks, and use copy of JRuby repo
I am using the personal copies of those repos just to have more control, and ensure that revisions do not get lost.
In each of the repos, I am going to add a tag for the revision that corresponds to the paper results. Just to make sure that we can find back the bits and pieces at a later point.

Signed-off-by: Stefan Marr <git@stefan-marr.de>