Added data file
Data file is added as bzip2 compressed version, because the plain text file is 17MB. (bzip2 gives better results than gzip).
We need to make sure that the data file stays consistent with the setup though.

Generally, the data file is just a 'cache', and to be considered out of date. The authoritative source should always be the benchmark results from ones one runs.

And, just to mention it: compression is handled transparently in R, so the file doesn't need to be decompressed manually.

Signed-off-by: Stefan Marr <git@stefan-marr.de>