Added minimal library of basic data structures to be used by the benchmarks
This contains only the bare minimum to make the benchmarks run. Nothing more. Might need to add a little more for other benchmarks.
And, if Java does not want to compile, we might need to fill in more blanks as well. Or remove code, and replace it with exceptions.

Note, the sort is not actually exercised in DeltaBlue, because there are never multiple elements.

Signed-off-by: Stefan Marr <git@stefan-marr.de>