Fix CD benchmark for Smalltalk.
The additional code in the Smalltalk version of `CollisionDetector>>isInVoxel:motion:` that forces IEEE floating point semantics used `Float negativeInfinity` instead of `Float infinity` for `low_x` and `low_y`. This has accidentally caused a max stack height of more than 500.000 (as opposed to 5), which in turn caused stack overflow errors in TruffleSqueak. With this fix in place, we can finally run the CD benchmark on TruffleSqueak.