[Ruby] Apply `# frozen_string_literal: true` to all benchmarks
It seems good engineering practice to have immutable strings, and it may enable further optimizations.

Arguably, this may even be more comparable to other languages, since strings are often immutable.

On the other hand, one may want a VM to figure out automatically that things are immutable to benefit from it. So, there’s a tradeoff here.

For the moment, it seems useful to go with the immutable version, which may become the default in Ruby at some point, too.

Signed-off-by: Stefan Marr <git@stefan-marr.de>